wip: Allow wakelock to timeout naturally
When an intent is sent off after a notification event is received, the
receiver may not obtain their own wakelock.  In that event, despite the
near-instant delivery of a push notification, the event receiver may yet
still need to wait for the phone to wake again in order to actually
act on the notification.

In this patch, we employ the existing
NOTIFICATION_RECEIVED_WAKELOCK_TIMEOUT_MILLIS value that controls the
length of time the wakelock is held before expiring.  Normally, this
would never occur, instead the lock is released at the end of our
processing of the event.  However, when we release it, the receiving
applications lose their ability to act before the device is accidentally
woken for some other reason.  Instead, we set this timeout to a more
reasonable value of 3 seconds, and do NOT release it ourselves.  This
means that the receiving applications have around 3 seconds to act
before the wakelock expires.

This is less than ideal: the receiving applications should know whether
or not their response deserves a wakelock, and they should themselves
get and hold it until they are satisfied.  However, this works around
buggy AOSP versions and applications that do not implement this more
ideal behavior.

Signed-off-by: Antonio Russo <aerusso@aerusso.net>