feat: Show icon of registered apps and optional message
UnifiedPush allows the end user app to send through an optional
message for the distributor to display. Use that, if present (and if the
user hasn't set an explicit display name) as the `displayName` for the
app.

Instead of showing a generic icon for the app, use the app's actual
icon.

Together, this make it more obvious to the user which app is responsible
for a particular registration, and the account (or other metadata) the
app is using for the registration.

Fixes #1166